+++ packages/api/src/entities/user-shops/index.ts
@@ -1,3 +1,0 @@
-export * from './user-shops.controller.js';
-export * from './user-shops.module.js';
-export * from './user-shops.service.js';