+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sales-planner/react
 
+## 0.4.0
+
+### Minor Changes
+
+- Remove user-shops entity (unused for access control, superseded by user-roles)
+
+### Patch Changes
+
+- Updated dependencies
+  - @sales-planner/shared@0.19.0
+  - @sales-planner/http-client@0.21.0
+
 ## 0.3.0
 
 ### Minor Changes